% Generated by build script
\name{stooq_forex}
\alias{stooq_forex}
\title{Fetch forex pair history}
\usage{
stooq_forex()
}
\arguments{
\item{pair}{Forex pair (e.g. "eurusd", "gbpusd", "usdjpy")}
\item{start}{Start date}
\item{end}{End date}
\item{interval}{"d", "w", or "m"}
}
\value{tibble: date, open, high, low, close, volume, ticker}
\description{Fetch forex pair history}
