% Generated by build script
\name{stooq_index}
\alias{stooq_index}
\title{Fetch US index history}
\usage{
stooq_index()
}
\arguments{
\item{index}{Index name: "spx" (S&P 500), "dji" (Dow Jones),}
\item{start}{Start date}
\item{end}{End date}
\item{interval}{"d", "w", or "m"}
}
\value{tibble: date, open, high, low, close, volume, ticker}
\description{Fetch US index history}
