% Generated by build script
\name{stooq_yield}
\alias{stooq_yield}
\title{Fetch US Treasury yield history}
\usage{
stooq_yield()
}
\arguments{
\item{maturity}{Maturity: "2" (2Y), "5" (5Y), "10" (10Y), "30" (30Y)}
\item{start}{Start date}
\item{end}{End date}
\item{interval}{"d", "w", or "m"}
}
\value{tibble: date, open, high, low, close, volume, ticker}
\description{Fetch US Treasury yield history}
